+++ Tests/ChessCoreTests/Rules/Standard/PawnMovesTests.swift
@@ -37,7 +37,7 @@ import Testing
         ("8/8/3p2k1/P2Pr1Pp/3R2KP/8/8/8 w - h6 0 2", "g5h6", "g5"),
     ])
 func assert(fen: String, moves: String, coordinate: String) {
-    let position = FENSerializer().position(from: fen)
+    let position = try! FENSerializer().position(from: fen)
     let square = Square(coordinate: coordinate)
     let moves = moves.split(separator: " ").map { "\($0)" }
     let testMoves = StandardRules().legalMovesForPiece(at: square, in: position)